Fixed bug in declarations typing.
The name of a variable can shadow the name of its type when declared.
Idem, the name of a function can shadow the type of its return type.
Furthermore, the name of a parameter can shadow the name of types of
following parameters. Finally, a function or a procedure is declared in
the parent context, no in its one (allows shadowing the functionnal in
its declarations field).